+++ src/mailman/docs/NEWS.rst
@@ -7,6 +7,11 @@ Copyright (C) 1998-2022 by the Free Software Foundation, Inc.
 
 Here is a history of user visible changes to Mailman.
 
+.. _news-3.3.9:
+
+3.3.9
+=====
+
 
 .. _news-3.3.8: